src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pyhd8ed1ab_0
aiohttp=3.7.4.post0=py38h7f8727e_2
anyio=3.4.0=py38h578d9bd_0
argon2-cffi=20.1.0=py38h27cfd23_1
async-timeout=3.0.1=py_1000
async_generator=1.10=py_0
attrs=21.4.0=pyhd8ed1ab_0
autopep8=1.6.0=pyhd3eb1b0_0
babel=2.9.1=pyh44b312d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bleach=4.1.0=pyhd8ed1ab_0
blinker=1.4=py_1
bottleneck=1.3.2=py38heb32a55_1
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py38h497a2fe_1001
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.4=pyhd8ed1ab_0
celluloid=0.2.0=pypi_0
certifi=2021.10.8=py38h06a4308_2
cffi=1.14.6=py38h400218f_0
chardet=4.0.0=py38h578d9bd_2
charset-normalizer=2.0.9=pyhd8ed1ab_0
click=8.0.3=py38h578d9bd_1
cryptography=35.0.0=py38ha5dfef3_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyhc8e2a94_3
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.0=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
descartes=1.1.0=pyhd3eb1b0_4
entrypoints=0.3=pyhd8ed1ab_1003
expat=2.4.1=h2531618_2
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h5202010_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.3.3=pyh6c4a22f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
googledrivedownloader=0.4=pyhd3deb0d_1
grpcio=1.42.0=py38hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.1=pyhd3deb0d_0
importlib-metadata=4.10.0=py38h578d9bd_0
importlib_resources=5.4.0=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.6.0=py38he5a9106_0
ipython=7.30.1=py38h578d9bd_0
ipython_genutils=0.2.0=py_1
jedi=0.18.1=py38h578d9bd_0
jinja2=3.0.3=pyhd8ed1ab_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=4.3.3=pyhd8ed1ab_0
jupyter_client=7.1.0=pyhd3eb1b0_0
jupyter_core=4.9.1=py38h06a4308_0
jupyter_server=1.13.2=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.3.1=py38h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pyhd8ed1ab_0
markupsafe=2.0.1=py38h497a2fe_0
matplotlib=3.5.0=py38h06a4308_0
matplotlib-base=3.5.0=py38h3ed280b_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py38h497a2fe_1004
mizani=0.7.3=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=5.2.0=py38h7f8727e_2
munkres=1.1.4=py_0
nbclassic=0.3.4=pyhd8ed1ab_0
nbclient=0.5.9=pyhd8ed1ab_0
nbconvert=6.4.0=py38h578d9bd_0
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.4=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pyhd8ed1ab_1
notebook=6.4.6=pyha770c72_0
numexpr=2.7.3=py38h22e1b3c_1
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
oauthlib=3.1.1=pyhd8ed1ab_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
palettable=3.3.0=pyhd3eb1b0_0
pandas=1.3.4=py38h8c16a72_0
pandoc=2.16.2=h7f98852_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
patsy=0.5.2=py38h06a4308_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.4.0=py38h5aabda8_0
pip=21.2.4=py38h06a4308_0
plotnine=0.8.0=pyhd8ed1ab_0
prometheus_client=0.12.0=pyhd8ed1ab_0
prompt-toolkit=3.0.24=pyha770c72_0
protobuf=3.17.2=py38h295c915_0
ptyprocess=0.7.0=pyhd3deb0d_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycodestyle=2.8.0=pyhd3eb1b0_0
pycparser=2.21=pyhd8ed1ab_0
pyg=2.0.3=py38_torch_1.10.0_cu113
pygments=2.10.0=pyhd8ed1ab_0
pyjwt=2.3.0=pyhd8ed1ab_1
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.18.0=py38heee7806_0
pysocks=1.7.1=py38h578d9bd_4
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-louvain=0.15=pyhd3deb0d_0
python_abi=3.8=2_cp38
pytorch=1.10.1=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-cluster=1.5.9=py38_torch_1.10.0_cu113
pytorch-mutex=1.0=cuda
pytorch-scatter=2.0.9=py38_torch_1.10.0_cu113
pytorch-sparse=0.6.12=py38_torch_1.10.0_cu113
pytorch-spline-conv=1.2.1=py38_torch_1.10.0_cu113
pytz=2021.3=pyhd3eb1b0_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=5.4.1=py38h497a2fe_0
pyzmq=19.0.2=py38ha71036d_2
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.26.0=pyhd8ed1ab_1
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.8=pyhd8ed1ab_0
scikit-learn=1.0.1=py38h51133e4_0
scipy=1.7.1=py38h292c36d_2
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=58.0.4=py38h06a4308_0
sip=4.19.13=py38he6710b0_0
six=1.16.0=pyhd3eb1b0_0
sniffio=1.2.0=py38h578d9bd_2
sqlite=3.36.0=hc218d9a_0
statsmodels=0.13.0=py38h7f8727e_0
tensorboard=2.7.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.0=py38h2b97feb_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorboardx=2.2=pyhd3eb1b0_0
terminado=0.12.1=py38h578d9bd_1
testpath=0.5.0=pyhd8ed1ab_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.10.1=py38_cu113
torchvision=0.11.2=py38_cu113
tornado=6.1=py38h497a2fe_1
tqdm=4.62.3=pyhd3eb1b0_1
traitlets=5.1.1=pyhd8ed1ab_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=1.2.3=pyhd8ed1ab_0
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
yacs=0.1.6=py_0
yaml=0.2.5=h516909a_0
yarl=1.6.3=py38h497a2fe_2
zeromq=4.3.4=h9c3ff4c_0
zipp=3.7.0=pyhd8ed1ab_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
